requirements.txt
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-text-splitters>=0.3.0
langchain-google-genai>=2.0.0
langgraph>=0.2.0
faiss-cpu>=1.8.0
sentence-transformers>=3.0.0
google-generativeai>=0.8.0
streamlit>=1.36.0
openpyxl>=3.1.0
python-docx>=1.1.0
pandas>=2.2.0